+++ sensor.h
@@ -7,6 +7,6 @@ struct sensor {
     char    *temp_path;
 };
 
-int sensor_poll (struct sensor *s);
+float sensor_poll (struct sensor *s);
 struct sensor *sensor_create (const char *hwmon_path, int index, int offset);
 void sensor_destroy (struct sensor *s);
